Rename asset localization tests
