+++ webapp/src/test/java/com/box/l10n/mojito/service/tm/localizeasset/LocalizeJsonAssetTest.java
@@ -1,4 +1,4 @@
-package com.box.l10n.mojito.service.tm.filetype;
+package com.box.l10n.mojito.service.tm.localizeasset;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
@@ -12,9 +12,9 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TMServiceJsonTest extends TMServiceFileTypeTestBase {
+public class LocalizeJsonAssetTest extends LocalizeAssetTestBase {
 
-  static Logger logger = LoggerFactory.getLogger(TMServiceJsonTest.class);
+  static Logger logger = LoggerFactory.getLogger(LocalizeJsonAssetTest.class);
 
   @Test
   public void testLocalizeJsonWithComments() throws Exception {
