+++ webapp/src/test/java/com/box/l10n/mojito/service/tm/localizeasset/LocalizePoAssetTest.java
@@ -1,4 +1,4 @@
-package com.box.l10n.mojito.service.tm.filetype;
+package com.box.l10n.mojito.service.tm.localizeasset;
 
 import static org.junit.Assert.assertEquals;
 
@@ -13,9 +13,9 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TMServicePoTest extends TMServiceFileTypeTestBase {
+public class LocalizePoAssetTest extends LocalizeAssetTestBase {
 
-  static Logger logger = LoggerFactory.getLogger(TMServicePoTest.class);
+  static Logger logger = LoggerFactory.getLogger(LocalizePoAssetTest.class);
 
   @Test
   public void testLocalizePoPluralJp() throws Exception {
