+++ webapp/src/test/java/com/box/l10n/mojito/service/tm/localizeasset/LocalizeTsAssetTest.java
@@ -1,4 +1,4 @@
-package com.box.l10n.mojito.service.tm.filetype;
+package com.box.l10n.mojito.service.tm.localizeasset;
 
 import static org.junit.Assert.assertEquals;
 
@@ -10,9 +10,9 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TMServiceTsTest extends TMServiceFileTypeTestBase {
+public class LocalizeTsAssetTest extends LocalizeAssetTestBase {
 
-  static Logger logger = LoggerFactory.getLogger(TMServiceTsTest.class);
+  static Logger logger = LoggerFactory.getLogger(LocalizeTsAssetTest.class);
 
   @Test
   public void testLocalizeTSFile() throws Exception {
