+++ webapp/src/test/java/com/box/l10n/mojito/service/tm/localizeasset/LocalizeXtbAssetTest.java
@@ -1,4 +1,4 @@
-package com.box.l10n.mojito.service.tm.filetype;
+package com.box.l10n.mojito.service.tm.localizeasset;
 
 import static org.junit.Assert.assertEquals;
 
@@ -12,9 +12,9 @@ import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class TMServiceXtbTest extends TMServiceFileTypeTestBase {
+public class LocalizeXtbAssetTest extends LocalizeAssetTestBase {
 
-  static Logger logger = LoggerFactory.getLogger(TMServiceXtbTest.class);
+  static Logger logger = LoggerFactory.getLogger(LocalizeXtbAssetTest.class);
 
   @Test
   public void testLocalizeXtb() throws Exception {
